Style edit of 0-print_list.c/0x12-linked_list project

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -3,12 +3,12 @@
 
 /**
  * print_list - prints all the elements of a list_t list.
- * @h : given list_t struct
+ * @h : given list_t list
  * Return: the number of nodes
  */
 size_t print_list(const list_t *h)
 {
-	int i = 0;
+	size_t i = 0;
 	/*const list_t *tmp = h;*/
 
 	while (h != NULL)
